fix: cleanup Resolves: OKTA-12345
